refactor(components.ovhfilereader): move to webUniverseComponents
ref: OM-166
