+++ client/app/components/fileEditor/file-editor.html
@@ -1,7 +1,7 @@
 <div class="file-editor">
     <div class="file-editor-toolbar">
         <div class="d-inline-block">
-            <input type="file" id="fileEditorUpload" data-ovh-filereader
+            <input type="file" id="fileEditorUpload" data-wuc-ovh-filereader
                    data-ng-model="fileModel.value">
             <label class="btn btn-icon" for="fileEditorUpload" role="button">
                 <span class="fa fa-upload" aria-hidden="true"></span>
