+++ client/app/webUniverseComponents/index.js
@@ -4,6 +4,7 @@ import wucConverter from './converter';
 import wucCron from './cron';
 import wucCronValidator from './cron-validator';
 import wucFileChange from './fileChange';
+import wucOvhFileReader from './ovhFileReader';
 import wucProgressBarElementCounter from './progressBarElementCounter';
 
 const moduleName = 'webUniverseComponents';
@@ -14,6 +15,7 @@ angular
     wucCron,
     wucCronValidator,
     wucFileChange,
+    wucOvhFileReader,
     wucProgressBarElementCounter,
   ]);
